requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via starlette
argon2-cffi==25.1.0
    # via -r requirements.in
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
bcrypt==4.0.1
    # via
    #   -r requirements.in
    #   passlib
certifi==2025.11.12
    # via requests
cffi==2.0.0
    # via argon2-cffi-bindings
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fastapi==0.127.1
    # via -r requirements.in
fonttools==4.61.1
    # via matplotlib
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
kiwisolver==1.4.9
    # via matplotlib
loguru==0.7.3
    # via -r requirements.in
matplotlib==3.10.8
    # via
    #   -r requirements.in
    #   seaborn
numpy==2.4.0
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   seaborn
packaging==25.0
    # via matplotlib
pandas==2.3.3
    # via
    #   -r requirements.in
    #   seaborn
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pillow==12.0.0
    # via matplotlib
psutil==7.2.0
    # via -r requirements.in
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pyotp==2.9.0
    # via -r requirements.in
pyparsing==3.3.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via pydantic-settings
pytz==2025.2
    # via pandas
requests==2.32.5
    # via -r requirements.in
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.45
    # via sqlmodel
sqlmodel==0.0.30
    # via -r requirements.in
starlette==0.50.0
    # via fastapi
tqdm==4.67.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
urllib3==2.6.2
    # via requests
uvicorn==0.40.0
    # via -r requirements.in
zxcvbn-python==4.4.24
    # via -r requirements.in
